added NavBar component with routes
